Fix LoadCollector throwing an IndexError
Some behavior changed in Numpy or Pint and it was causing the load
collector to throw an IndexError when checking if the contained type was
array like. This commit fixes the issue by catching the IndexError and
treating it as an indication that the contained type is not array like.